refactor(#334): 영수증 컴포넌트 공통화
